+++ scripts/cy-wrapper.sh
@@ -3,6 +3,11 @@
 # Warning: this script is expected to be used in cycloid-toolkit docker image.
 # We do not recommand to use it in another context.
 
+if [ -n "$CY_DEBUG" ]; then
+  echo "CY_DEBUG provided, wrapper running in DEBUG mode" >&2
+  set -x
+fi
+
 if test -z "$BASH_VERSION"; then
   echo "Please run this script using bash, not sh or any other shell." >&2
   exit 1